src/requirements.txt
fastapi==0.59.0
uvicorn==0.11.6
requests==2.24.0
asyncpg==0.20.1
sqlalchemy==1.3.18
databases==0.3.2
boto3==1.14.23
python-decouple==3.3
psycopg2-binary==2.8.5